refactoring-1 abstract class를 앞단에 배치하여 공통적인 로직(service 메소드) 를 구축하고 추상 메소드만 각 핸들러에서 구현하도록 수정
