+++ src/main/java/handler/HttpHandler.java
@@ -1,0 +1,12 @@
+package handler;
+
+import webserver.HttpRequest;
+import webserver.HttpResponse;
+
+public interface HttpHandler {
+    HttpResponse service(HttpRequest httpRequest);
+}
+
+class InvalidUrlException extends RuntimeException {
+
+}
