+++ src/main/java/handler/UserHandler.java
@@ -1,4 +1,4 @@
-package webserver;
+package handler;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -13,9 +13,11 @@ import org.slf4j.LoggerFactory;
 import model.User;
 import service.UserService;
 import util.HttpRequestUtils;
-import util.HttpResponseUtils;
+import webserver.HttpRequest;
+import webserver.HttpResponse;
+import webserver.HttpStatusCode;
 
-public class UserHandler implements HttpHandler {
+public class UserHandler extends AbstracrtHandler {
     private static final Logger log = LoggerFactory.getLogger(ResourceHandler.class);
     //이거 나중에 통일되도록 뽑자
     private static final String WEB_RESOURCE_ROOT = "/Users/kakao/workspace/web-application-server/webapp";
@@ -26,18 +28,30 @@ public class UserHandler implements HttpHandler {
     }
 
     @Override
-    public HttpResponse service(HttpRequest httpRequest) {
-        String resource = httpRequest.getResource();
+    public HttpResponse doPost(HttpRequest httpRequest) {
         HttpResponse response = null;
+        String resource = httpRequest.getResource();
 
         if("/user/create".equals(resource)) {
             response = signUp(httpRequest.getBody());
         } else if ("/user/login".equals(resource)) {
             response = login(httpRequest.getBody());
-        } else if ("/user/list".equals(resource)) {
+        } else {
+            throw new InvalidUrlException();
+        }
+
+        return response;
+    }
+
+    @Override
+    public HttpResponse doGet(HttpRequest httpRequest) {
+        HttpResponse response;
+        String resource = httpRequest.getResource();
+
+        if ("/user/list".equals(resource)) {
             response = getUserList(httpRequest);
         } else {
-            response = HttpResponseUtils.make_404_response();
+            throw new InvalidUrlException();
         }
 
         return response;
@@ -66,11 +80,11 @@ public class UserHandler implements HttpHandler {
         if(user != null) {
             headers.put("Location", "/index.html");
             headers.put("Set-Cookie", "logined=true");
-            return new HttpResponse(HttpStatusCode.REDIRECT, headers,null);
         } else {
             headers.put("Location", "/user/login_failed.html");
-            return new HttpResponse(HttpStatusCode.REDIRECT, headers,null);
         }
+
+        return new HttpResponse(HttpStatusCode.REDIRECT, headers,null);
     }
 
     private HttpResponse getUserList(HttpRequest httpRequest) {
