--- src/main/java/webserver/HttpHandler.java
@@ -1,5 +1,0 @@
-package webserver;
-
-public interface HttpHandler {
-    HttpResponse service(HttpRequest httpRequest);
-}
